Replace regex matching with simple object lookup
This is a lot more effecient way to find whitelist matches. We do not
have to loop over all fields and use regex, instead we can use a simple
hashmap lookup.

This commit also removed matches for fields with tildes in it. They are
exceptions and we haven't seen them being used.